requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
jinja2==3.1.4
pydantic==2.9.2
faiss-cpu==1.8.0.post1
sentence-transformers==3.2.0
numpy==1.26.4
scikit-learn==1.5.2
requests==2.32.3
pytest==8.3.3
httpx==0.27.2
